requirements.txt
# Web Scraping
beautifulsoup4==4.12.2
requests==2.31.0
playwright==1.40.0
selenium==4.16.0

# Semantic Matching (NEW - for smart_watcher_v2)
sentence-transformers==2.2.2
numpy==1.24.3
torch==2.1.0  # Required by sentence-transformers

# LLM Integration
openai==1.6.1
anthropic==0.8.1
google-genai>=0.2.0

# PDF Generation
Jinja2==3.1.2
pypandoc==1.12

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
discord-webhook==1.3.0

# Dashboard
streamlit==1.29.0

# Data Storage & Analytics
pandas==2.1.4
